fix: protect pages
