+++ app/about/layout.tsx
@@ -1,16 +1,22 @@
 import { Metadata } from 'next';
+import { cookies } from 'next/headers';
+import { redirect } from 'next/navigation';
 
 export const metadata: Metadata = {
   title: 'CompanyScore | О нас',
   description: 'Generated by create next app',
 };
 
-export default function CompaniesLayout({
+export default async function AboutLayout({
   children,
 }: Readonly<{
   children: React.ReactNode;
 }>) {
-  return (
-    <main className="flex-1 container mx-auto h-full px-4">{children}</main>
-  );
+  const cookieStore = await cookies();
+  const accessToken = cookieStore.get('accessToken')?.value;
+
+  if (!accessToken) {
+    redirect('/login');
+  }
+  return <>{children}</>;
 }
