+++ app/layout.tsx
@@ -1,6 +1,3 @@
-import type { Metadata } from 'next';
-import { cookies } from 'next/headers';
-import { redirect } from 'next/navigation';
 import { Inter } from 'next/font/google';
 import { Header, HeaderMobile, Footer } from '@/widgets';
 import './globals.css';
@@ -11,23 +8,11 @@ const inter = Inter({
   variable: '--font-inter',
 });
 
-export const metadata: Metadata = {
-  title: 'CompanyScore | Главная',
-  description: 'Generated by create next app',
-};
-
 export default async function RootLayout({
   children,
 }: Readonly<{
   children: React.ReactNode;
 }>) {
-  const cookieStore = await cookies();
-  const accessToken = cookieStore.get('accessToken')?.value;
-
-  if (!accessToken) {
-    redirect('/login');
-  }
-
   return (
     <html lang="en">
       <body
