+++ app/page.tsx
@@ -1,12 +1,21 @@
 import { Metadata } from 'next';
+import { cookies } from 'next/headers';
+import { redirect } from 'next/navigation';
 import { Hero, Features, Steps, Reviews, Blog, Call } from './components';
 
 export const metadata: Metadata = {
-  title: 'CompanyScore | Home',
+  title: 'CompanyScore | Главная',
   description: 'Generated by create next app',
 };
 
-export default function HomePage() {
+export default async function HomePage() {
+  const cookieStore = await cookies();
+  const accessToken = cookieStore.get('accessToken')?.value;
+
+  if (!accessToken) {
+    redirect('/login');
+  }
+
   return (
     <section className="flex flex-col items-stretch justify-center">
       <Hero />
